Сделана модуляция
