+++ js/varibles.js
@@ -1,0 +1,24 @@
+const NUMBER_OF_POSTS = 25;
+const MIN_NUMBERS_OF_LIKES = 15;
+const MAX_NUMBERS_OF_LIKES = 200;
+const MAX_COMMENTS = 100;
+const MIN_AVATAR_ID = 1;
+const MAX_AVATAR_ID = 6;
+const descriptions = ['Отличный вид', 'Вернуться бы еще раз сюда', 'Я это съел', 'Мои друзья', 'Посиделки'];
+const messages = ['В целом всё неплохо. Но не всё.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.'];
+const names = ['Михаил', 'Иван', 'Сергей', 'Татьяна', 'Мария', 'Ольга', 'Ярослава'];
+const COUNTER = 1;
+
+export {
+  NUMBER_OF_POSTS,
+  MIN_NUMBERS_OF_LIKES,
+  MAX_NUMBERS_OF_LIKES,
+  MAX_COMMENTS,
+  MIN_AVATAR_ID,
+  MAX_AVATAR_ID,
+  descriptions,
+  messages,
+  names,
+  COUNTER,
+};
